requirements.txt
tweepy==4.14.0
python-dotenv==1.0.0
openai==0.28.1